# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Cidade Aberta - Plataforma de Participação Cidadã
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /demands:
    get:
      operationId: listDemands
      tags: [demands]
      summary: List public demands
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: categoryId
          in: query
          schema:
            type: ["integer", "null"]
        - name: neighborhoodId
          in: query
          schema:
            type: ["integer", "null"]
        - name: departmentId
          in: query
          schema:
            type: ["integer", "null"]
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of demands
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DemandListResponse"

    post:
      operationId: createDemand
      tags: [demands]
      summary: Create a new demand
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DemandInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Demand"
  /demands/{id}:
    get:
      operationId: getDemand
      tags: [demands]
      summary: Get demand details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Demand
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DemandDetail"
        "404":
          description: Not found

    patch:
      operationId: updateDemand
      tags: [demands]
      summary: Update demand (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DemandUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Demand"
  /demands/{id}/timeline:
    get:
      operationId: getDemandTimeline
      tags: [demands]
      summary: Get public timeline of a demand
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Timeline events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimelineEvent"
  /demands/{id}/status:
    patch:
      operationId: updateDemandStatus
      tags: [demands]
      summary: Update demand status (admin/secretary)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Demand"

  # COMMENTS
...
  /departments/{id}/stats:
    get:
      operationId: getDepartmentStats
      tags: [departments]
      summary: Get stats for a department
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Department statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepartmentStats"

  # PUBLIC WORKS
  /public-works:
    get:
      operationId: listPublicWorks
      tags: [public-works]
      summary: List public works and projects
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Public works
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PublicWork"

    post:
      operationId: createPublicWork
      tags: [public-works]
      summary: Create public work (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublicWorkInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicWork"
  /public-works/{id}:
    get:
      operationId: getPublicWork
      tags: [public-works]
      summary: Get public work details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Public work
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicWork"

  # DASHBOARD
...
components:
  schemas:
...
    Demand:
      type: object
      required: [id, title, description, status, categoryId, neighborhoodId, address, lat, lng, confirmationsCount, commentsCount, createdAt, authorId]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [open, under_review, approved, in_progress, completed, rejected, closed]
        priority:
          type: ["string", "null"]
          enum: [low, medium, high, urgent, null]
        categoryId:
          type: integer
        category:
          $ref: "#/components/schemas/Category"
        neighborhoodId:
          type: integer
        neighborhood:
          $ref: "#/components/schemas/Neighborhood"
        departmentId:
          type: ["integer", "null"]
        address:
          type: string
        lat:
          type: number
        lng:
          type: number
        photoUrl:
          type: ["string", "null"]
        confirmationsCount:
          type: integer
        commentsCount:
          type: integer
        isConfirmedByMe:
          type: boolean
        authorId:
          type: integer
        authorName:
          type: string
        publicResponse:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    DemandDetail:
      type: object
      required: [id, title, description, status, categoryId, neighborhoodId, address, lat, lng, confirmationsCount, commentsCount, createdAt, authorId, timeline]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [open, under_review, approved, in_progress, completed, rejected, closed]
        priority:
          type: ["string", "null"]
          enum: [low, medium, high, urgent, null]
        categoryId:
          type: integer
        category:
          $ref: "#/components/schemas/Category"
        neighborhoodId:
          type: integer
        neighborhood:
          $ref: "#/components/schemas/Neighborhood"
        departmentId:
          type: ["integer", "null"]
        department:
          $ref: "#/components/schemas/Department"
        address:
          type: string
        lat:
          type: number
        lng:
          type: number
        photoUrl:
          type: ["string", "null"]
        confirmationsCount:
          type: integer
        commentsCount:
          type: integer
        isConfirmedByMe:
          type: boolean
        authorId:
          type: integer
        authorName:
          type: string
        publicResponse:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
        timeline:
          type: array
          items:
            $ref: "#/components/schemas/TimelineEvent"
        serviceOrders:
          type: array
          items:
            $ref: "#/components/schemas/ServiceOrder"
    DemandListResponse:
      type: object
      required: [demands, total]
      properties:
        demands:
          type: array
          items:
            $ref: "#/components/schemas/Demand"
        total:
          type: integer
    DemandInput:
      type: object
      required: [title, description, categoryId, neighborhoodId, address, lat, lng]
      properties:
        title:
          type: string
          minLength: 5
        description:
          type: string
          minLength: 10
        categoryId:
          type: integer
        neighborhoodId:
          type: integer
        address:
          type: string
        lat:
          type: number
        lng:
          type: number
        photoUrl:
          type: ["string", "null"]
    DemandUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        priority:
          type: ["string", "null"]
        departmentId:
          type: ["integer", "null"]
        publicResponse:
          type: ["string", "null"]
    StatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [open, under_review, approved, in_progress, completed, rejected, closed]
        note:
          type: ["string", "null"]
        publicResponse:
          type: ["string", "null"]
    TimelineEvent:
      type: object
      required: [id, demandId, event, createdAt]
      properties:
        id:
          type: integer
        demandId:
          type: integer
        event:
          type: string
        description:
          type: ["string", "null"]
        actorName:
          type: ["string", "null"]
        actorRole:
          type: ["string", "null"]
        statusFrom:
          type: ["string", "null"]
        statusTo:
          type: ["string", "null"]
        isPublic:
          type: boolean
        createdAt:
          type: string

    # COMMENT
...
    Category:
      type: object
      required: [id, name, icon, color]
      properties:
        id:
          type: integer
        name:
          type: string
        icon:
          type: string
        color:
          type: string
        description:
          type: ["string", "null"]
        demandsCount:
          type: integer
...
    Neighborhood:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        demandsCount:
          type: integer
        resolvedCount:
          type: integer

    # SERVICE ORDER
    ServiceOrder:
      type: object
      required: [id, demandId, status, createdAt]
      properties:
        id:
          type: integer
        demandId:
          type: integer
        demand:
          $ref: "#/components/schemas/Demand"
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, assigned, in_progress, completed, cancelled]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        teamId:
          type: ["integer", "null"]
        team:
          $ref: "#/components/schemas/Team"
        estimatedCompletion:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        slaHours:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
...
    Team:
      type: object
      required: [id, name, area]
      properties:
        id:
          type: integer
        name:
          type: string
        area:
          type: string
        description:
          type: ["string", "null"]
        activeOrdersCount:
          type: integer
        completedOrdersCount:
          type: integer
...
    Department:
      type: object
      required: [id, name, description]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        logoUrl:
          type: ["string", "null"]
        demandsCount:
          type: integer
        resolvedCount:
          type: integer
        avgResolutionDays:
          type: ["number", "null"]
        efficiencyIndex:
          type: ["number", "null"]
    DepartmentStats:
      type: object
      required: [departmentId, totalDemands, resolved, inProgress, avgResolutionDays]
      properties:
        departmentId:
          type: integer
        totalDemands:
          type: integer
        resolved:
          type: integer
        inProgress:
          type: integer
        rejected:
          type: integer
        avgResolutionDays:
          type: number
        efficiencyIndex:
          type: number
        recentDemands:
          type: array
          items:
            $ref: "#/components/schemas/Demand"

    # PUBLIC WORKS
    PublicWork:
      type: object
      required: [id, title, description, status, neighborhoodId, startDate, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [planned, approved, in_progress, completed, suspended]
        neighborhoodId:
          type: integer
        neighborhood:
          $ref: "#/components/schemas/Neighborhood"
        departmentId:
          type: ["integer", "null"]
        investment:
          type: ["number", "null"]
        startDate:
          type: ["string", "null"]
        estimatedEnd:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        lat:
          type: ["number", "null"]
        lng:
          type: ["number", "null"]
        photoUrl:
          type: ["string", "null"]
        progressPercent:
          type: integer
        createdAt:
          type: string
    PublicWorkInput:
      type: object
      required: [title, description, neighborhoodId, status]
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [planned, approved, in_progress, completed, suspended]
        neighborhoodId:
          type: integer
        departmentId:
          type: ["integer", "null"]
        investment:
          type: ["number", "null"]
        startDate:
          type: ["string", "null"]
        estimatedEnd:
          type: ["string", "null"]
        lat:
          type: ["number", "null"]
        lng:
          type: ["number", "null"]
        progressPercent:
          type: integer

    # DASHBOARD
...