# lib/api-zod/src/generated/types/publicWork.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Cidade Aberta - Plataforma de Participação Cidadã
 * OpenAPI spec version: 0.1.0
 */
import type { Neighborhood } from './neighborhood';
import type { PublicWorkStatus } from './publicWorkStatus';

export interface PublicWork {
  id: number;
  title: string;
  description: string;
  status: PublicWorkStatus;
  neighborhoodId: number;
  neighborhood?: Neighborhood;
  /** @nullable */
  departmentId?: number | null;
  /** @nullable */
  investment?: number | null;
  /** @nullable */
  startDate: string | null;
  /** @nullable */
  estimatedEnd?: string | null;
  /** @nullable */
  completedAt?: string | null;
  /** @nullable */
  lat?: number | null;
  /** @nullable */
  lng?: number | null;
  /** @nullable */
  photoUrl?: string | null;
  progressPercent?: number;
  createdAt: string;
}